@@ -3,39 +3,41 @@
   "version": "1.0.0",
   "scripts": {
     "build": "webpack",
-    "start": "webpack serve --open",
     "lint": "eslint '**/*.ts'",
-    "clean": "gts clean",
+    "lint:fix": "eslint '**/*.ts' --fix",
+    "clean": "rm -rf dist",
     "commitlint": "commitlint --from=HEAD~1 --to=HEAD",
     "compile": "tsc",
-    "fix": "gts fix",
     "prepare": "husky",
     "pretest": "npm run compile",
-    "posttest": "npm run lint || true", 
+    "posttest": "npm run lint || true",
     "test": "echo \"No tests specified\" && exit 0"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.6.1",
     "@commitlint/config-conventional": "^19.6.0",
-    "@types/node": "^22.2.0",
-    "@typescript-eslint/eslint-plugin": "^8.21.0",
-    "@typescript-eslint/parser": "^8.21.0",
-    "esbuild": "^0.23.0",
-    "esbuild-plugin-tsc": "^0.4.0",
-    "eslint": "^8.57.1",
-    "gts": "^5.3.1",
+    "@eslint/js": "^9.19.0",
+    "@types/eslint": "^9.6.1",
+    "@types/node": "^22.12.0",
+    "@types/webpack": "^5.28.5",
+    "@typescript-eslint/eslint-plugin": "^8.22.0",
+    "@typescript-eslint/parser": "^8.22.0",
+    "eslint": "^9.19.0",
+    "eslint-import-resolver-typescript": "^3.7.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-prettier": "^5.2.3",
     "husky": "^9.1.7",
-    "ts-loader": "^9.5.1",
+    "ts-loader": "^9.5.2",
     "ts-node": "^10.9.2",
-    "tsup": "^8.2.4",
     "typescript": "^5.7.3",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "typescript-eslint": "^8.22.0",
+    "webpack": "^5.97.1",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.0"
   },
   "dependencies": {
-    "gts": "^5.3.1",
-    "hotkeys-js": "^3.13.7",
-    "zod": "^3.23.8"
+    "hotkeys-js": "^3.13.9",
+    "zod": "^3.24.1"
   }
 }
